+++ libft/ft_atoi.c
@@ -6,14 +6,14 @@
 /*   By: rficht <robin.ficht@free.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/03 16:46:07 by mdjemaa           #+#    #+#             */
-/*   Updated: 2023/04/29 15:25:27 by rficht           ###   ########.fr       */
+/*   Updated: 2023/06/01 10:55:49 by rficht           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../Includes/libft.h"
 #include <stdio.h>
 
-long	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int		signe;
 	size_t	nb;
